ENH: stats: Add an sf method to levy_l for improved precision. (#12757)
* ENH: stats: Add an sf method to levy_l for improved precision.

The CDF of `levy_l` is implemented as `2*F - 1`, where `F` is the output
of the CDF of the normal distribution evaluated at a function of x.  The
survival function is `1 - CDF = 1 - (2*F - 1) = 2*(1 - F) = 2*S`, where
`S` is the survival function of the normal distribution evaluated at the
same function of x.  By using that expression in the `_sf` method, we
avoid the subtraction that results in the catastrophic loss of precision.

* ENH: stats: Add an isf method to levy_l for improved precision.

Use an explicit formula for the inverse survival function to get better
precision for very small input values.